--- core/src/main/java/life/qbic/domain/datamanagement/Example.java
@@ -1,1 +1,0 @@
-
