+++ core/src/main/java/life/qbic/identityaccess/application/notifications/MessageBusInterface.java
@@ -1,4 +1,4 @@
-package life.qbic.apps.datamanager.notifications;
+package life.qbic.identityaccess.application.notifications;
 
 public interface MessageBusInterface {
