+++ core/src/main/java/life/qbic/identityaccess/application/notifications/MessageSubscriber.java
@@ -1,4 +1,4 @@
-package life.qbic.apps.datamanager.notifications;
+package life.qbic.identityaccess.application.notifications;
 
 /**
  * Interface for notification subscribers. This interface needs to be implemented, when a client
