+++ webapp/src/main/java/life/qbic/Application.java
@@ -10,17 +10,17 @@ import com.vaadin.flow.theme.Theme;
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
 import java.io.ObjectInputStream;
-import life.qbic.apps.datamanager.notifications.MessageBusInterface;
-import life.qbic.apps.datamanager.notifications.MessageSubscriber;
-import life.qbic.domain.usermanagement.DomainRegistry;
-import life.qbic.domain.usermanagement.registration.ConfirmEmailOutput;
-import life.qbic.domain.usermanagement.registration.EmailAddressConfirmation;
-import life.qbic.identityaccess.domain.user.UserDomainService;
-import life.qbic.domain.usermanagement.registration.UserRegistered;
-import life.qbic.domain.usermanagement.repository.UserRepository;
 import life.qbic.email.Email;
 import life.qbic.email.EmailService;
 import life.qbic.email.Recipient;
+import life.qbic.identityaccess.application.notifications.MessageBusInterface;
+import life.qbic.identityaccess.application.notifications.MessageSubscriber;
+import life.qbic.identityaccess.application.user.ConfirmEmailOutput;
+import life.qbic.identityaccess.application.user.EmailAddressConfirmation;
+import life.qbic.identityaccess.domain.events.DomainRegistry;
+import life.qbic.identityaccess.domain.user.UserDomainService;
+import life.qbic.identityaccess.domain.user.UserRegistered;
+import life.qbic.identityaccess.domain.user.UserRepository;
 import life.qbic.usermanagement.registration.EmailConfirmationLinkSupplier;
 import life.qbic.usermanagement.registration.EmailFactory;
 import life.qbic.views.login.LoginHandler;
