+++ webapp/src/main/java/life/qbic/events/EventRepository.java
@@ -1,7 +1,7 @@
 package life.qbic.events;
 
 import java.util.List;
-import life.qbic.domain.events.DomainEvent;
+import life.qbic.identityaccess.domain.events.DomainEvent;
 
 /**
  * A repository for domain events.
