+++ webapp/src/main/java/life/qbic/events/SimpleEventStore.java
@@ -3,7 +3,7 @@ package life.qbic.events;
 import java.util.Set;
 import java.util.stream.Collectors;
 import life.qbic.apps.datamanager.events.EventStore;
-import life.qbic.domain.events.DomainEvent;
+import life.qbic.identityaccess.domain.events.DomainEvent;
 
 /**
  * Implementation of a basic event store. It handles events and provides accessor methods to retain
