+++ webapp/src/main/java/life/qbic/usermanagement/persistence/UserJpaRepository.java
@@ -4,8 +4,8 @@ import java.util.List;
 import java.util.Optional;
 import life.qbic.identityaccess.domain.user.EmailAddress;
 import life.qbic.identityaccess.domain.user.User;
+import life.qbic.identityaccess.domain.user.UserDataStorage;
 import life.qbic.identityaccess.domain.user.UserId;
-import life.qbic.domain.usermanagement.repository.UserDataStorage;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
