+++ webapp/src/main/java/life/qbic/views/login/LoginHandler.java
@@ -4,8 +4,8 @@ import com.vaadin.flow.component.UI;
 import com.vaadin.flow.router.BeforeEvent;
 import java.util.List;
 import java.util.Map;
-import life.qbic.domain.usermanagement.registration.ConfirmEmailInput;
-import life.qbic.domain.usermanagement.registration.ConfirmEmailOutput;
+import life.qbic.identityaccess.application.user.ConfirmEmailInput;
+import life.qbic.identityaccess.application.user.ConfirmEmailOutput;
 import life.qbic.views.components.ErrorMessage;
 import life.qbic.views.components.InformationMessage;
 import org.springframework.beans.factory.annotation.Autowired;
